ok1

+++ REST-API/controllers/feed.js
@@ -1,6 +1,8 @@
 const { validationResult } = require('express-validator');
 const Post = require('../models/post');
 
+const io = require('../socket');
+
 const fs = require('fs');
 const path = require('path');
 const User = require('../models/user');
@@ -31,7 +33,7 @@ exports.getPosts = async (req, res, next) => {
   }
 };
 
-exports.createPost = (req, res, next) => {
+exports.createPost = async (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     const error = new Error('Filled data is incorrect!');
@@ -53,53 +55,54 @@ exports.createPost = (req, res, next) => {
     imageUrl: imageUrl,
     creator: req.userId
   });
-  post.save()
-    .then(result => {
-      return User.findById(req.userId);
-    })
-    .then(user => {  
-      creator = user;
-      user.posts.push(post);
-      return user.save();
+  try {
+        await post.save()
       
-    })
-    .then(result => {
-      res.status(201).json({
-        message: "Post created successfully!",
-        post: post,
-        creator: {_id: creator._id, name: creator.name}
-      });
-    })
-    .catch(err => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
-    })
+    const user = await User.findById(req.userId);
+   
+        creator = user;
+        user.posts.push(post);
+        await user.save();
+
+        io.getIO().emit('posts', { action: 'create', post: post })
+        
+        res.status(201).json({
+          message: "Post created successfully!",
+          post: post,
+          creator: {_id: creator._id, name: creator.name}
+        });
+  } catch(err) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+    next(err);
+  }
+
 
 };
 
-exports.getPost = (req, res, next) => {
+exports.getPost = async (req, res, next) => {
   const postId = req.params.postId;
-
-  Post.findById(postId)
-    .then(post => {
+  try { 
+  const post = await Post.findById(postId)
+                         .populate('creator')
+  
       if (!post) {
         const error = new Error("Couldn't Find the post");
         error.statusCode = 404;
         throw error;
       }
       res.status(200).json({ message: 'Post fetched', post: post })
-    })
-    .catch(err => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
-    })
+} catch(err) {
+  if (!err.statusCode) {
+    err.statusCode = 500;
+  }
+  next(err);
 }
 
-exports.updatePost = (req, res, next) => {
+}
+
+exports.updatePost = async (req, res, next) => {
   const postId = req.params.postId;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -118,41 +121,9 @@ exports.updatePost = (req, res, next) => {
     error.statusCode = 422;
     throw error;
   }
-  Post.findById(postId)
-    .then(post => {
-      if (!post) {
-        const error = new Error("Couldn't Find the post");
-        error.statusCode = 404;
-        throw error;
-      }
-      if(post.creator.toString() !== req.userId) {
-        const error = new Error('Not Authorized');
-        error.statusCode = 403;
-        throw error;
-      }
-      if(imageUrl != post.imageUrl){
-        clearImage(post.imageUrl);
-      }
-      post.title = title;
-      post.imageUrl = imageUrl;
-      post.content = content;
-      return post.save();
-    })
-    .then(result => {
-      res.status(200).json({message: 'Post Updated!!', post: result})
-    })
-    .catch(err => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
-    })
-}
-
-exports.deletePost = (req,res,next) => {
-  const postId = req.params.postId;
-  Post.findById(postId)
-      .then(post => {
+  try {
+    const post = await Post.findById(postId)
+      
         if (!post) {
           const error = new Error("Couldn't Find the post");
           error.statusCode = 404;
@@ -163,27 +134,59 @@ exports.deletePost = (req,res,next) => {
           error.statusCode = 403;
           throw error;
         }
-        clearImage(post.imageUrl);
-        return Post.findByIdAndRemove(postId);
-      })
-      .then(result => {
-        return User.findById(req.userId);
-        
-      })
-      .then(user => {
-        user.posts.pull(postId);
-        return user.save();
-        
-      })
-      .then(result => {
-        res.status(200).json({message: 'Post Deleted Successfully!!!'});
-      })
-      .catch(err => {
-        if (!err.statusCode) {
-          err.statusCode = 500;
+        if(imageUrl != post.imageUrl){
+          clearImage(post.imageUrl);
         }
-        next(err);
-      })
+        post.title = title;
+        post.imageUrl = imageUrl;
+        post.content = content;
+  
+  const result = await post.save();
+   
+  
+        res.status(200).json({message: 'Post Updated!!', post: result})
+  } catch(err) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+    next(err);
+  }
+
+}
+
+exports.deletePost = async (req,res,next) => {
+  const postId = req.params.postId;
+  try {
+    const post = await Post.findById(postId)
+  
+          if (!post) {
+            const error = new Error("Couldn't Find the post");
+            error.statusCode = 404;
+            throw error;
+          }
+          if(post.creator.toString() !== req.userId) {
+            const error = new Error('Not Authorized');
+            error.statusCode = 403;
+            throw error;
+          }
+          clearImage(post.imageUrl);
+          
+          await Post.findByIdAndRemove(postId);
+  
+          const user = await User.findById(req.userId);
+          
+          user.posts.pull(postId);
+  
+           await user.save();
+          
+          res.status(200).json({message: 'Post Deleted Successfully!!!'});
+  } catch(err) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+    next(err);
+  }
+
 }
 
 const clearImage = filePath => {
